requirements.txt
asgiref==3.7.2
contourpy==1.1.1
crispy==0.7.4
cycler==0.11.0
Django==4.2.5
django-cors-headers==4.3.0
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
fabio==2023.6.0
fonttools==4.42.1
h5py==3.9.0
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.7
PyJWT==1.7.1
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
